Add kontain specialization code to crun
This code adds the following:
- when crun is invoked with a symlink named krun that points to crun, crun will
  add bind mounts for /opt/kontain/runtime/libc.so, /opt/kontain/bin/km, and /dev/kvm (or /dev/kkm)
  to the config.json file for the container.  There is no need to add -v options on the krun
  command line to create these.
- when starting the entry point to the container, if the command being executed in not /opt/kontain/bin/km
  or a series of symlinks that resolve to km, then we insert /opt/kontain/bin/km into into the argument
  list before all of the existing command line arguments.
- There is some crude debug code in here that is not permanent.

It also allows some non km executables to run (by exec) in a kontain container

"docker exec ..." can run permitting non-kontain executables
in a kontain container.  A file containing a list of permitted executables is added to the image.
That file is scanned to see if the exec'ed program is in the list.  If it is, the program is run without
km being involved.  If the program is not in the list, it will be run with the assistance of km.

The allowed executables file is lines consisting of 3 fields separated by colons.
The fields are:
  1 - a regular expression that the desired executable should be matched with.
  2 - the name of the program to be run for that match.
  3 - the hash for the program in field 2.
When a matching entry is found, we verify that the program in field 2 has the hash in
field 3.  It the hash matches we return the program from field 2 to the caller requesting permission to
run the passed program.  If the hash does not match, krun terminates.
If no matching entry is found, no program path is returned to the caller.

Note:
When you specify km in the dockerfile ENTRYPOINT statement and run a container
with the --init flag you end up with a command line like this:

/opt/kontain/bin/km /sbin/docker-init -- /opt/kontain/bin/km ...the.payload.filename...

Since docker-init runs as a km payload when it execs to what you want running, it thinks
km is the payload.  That doesn't work.  This change removes the extra km argument when
/sbin/docker-init or /dev/init (podman) are present in argv[1].