Allow some non km executables to run (by exec) in a kontain container (#8)
* Allow some non km executables to run (by exec) in a kontain container

This change introduces the ability for "docker exec ..." to run permitted non-kontain executables
in a kontain container.  A file containing a list of permitted executables is added to the image.
That file is scanned to see if the exec'ed program is in the list.  If it is, the program is run without
km being involved.  If the program is not in the list, it will be run with the assistance of km.

The allowed executables file is lines consisting of 3 fields separated by colons.
The fields are:
  1 - a regular expression that the desired executable should be matched with.
  2 - the name of the program to be run for that match.
  3 - the hash for the program in field 2.
When a matching entry is found, we verify that the program in field 2 has the hash in
field 3.  It the hash matches we return the program from field 2 to the caller requesting permission to
run the passed program.  If the hash does not match, krun terminates.
If no matching entry is found, no program path is returned to the caller.

In addition 3 tests were added to the crun test_exec.py.  These tests are krun specific, they don't run
for the crun variety of the tests.