Removed 'symlink to km' traversal
After the latest PR (#1203, #1204) and discussions on symlink, we decided that
we need symlinks ONLY to bootstrap payloads in Linux, when we do not control executable start.
In KRUN when we fully control the start and ALWAYS add KM in front of the payload,
symlink traversal just breaks some use cases (e.g.  busybox, whihc uses symlinks for its own purposes)

So this fix removes symlink traversal support from KRUN

Tested manually on busybox runenv
We have an issue for adding KRUN-basd KM tests later