Allow /sbin/docker-init to run without km.
When a container is started with the --init flag, the first process in the container will
be /sbin/docker-init.  kontain does not supply or build this executable.  Currently krun
tries to run docker-init as a km payload which currently has some problems.  An expedient
solution is to have km allow docker-init run without km.  This change causes krun to run
docker-init without krun.  I'm not sure if this is the correct thing to do but it helps
get beyond this problem so we can test other things.