Change libcrun_kontain_argv() to remove extra km arguments. (#13)
When you specify km in the dockerfile ENTRYPOINT statement and run a container
with the --init flag you end up with a command line like this:

/opt/kontain/bin/km /sbin/docker-init -- /opt/kontain/bin/km ...the.payload.filename...

Since docker-init runs as a km payload when it execs to what you want running, it thinks
km is the payload.  That doesn't work.  This change removes the extra km argument when
/sbin/docker-init or /dev/init (podman) are present in argv[1].